Moving to adjacent line replaces previous position in history
This makes the plugin functionality more similar to jump lists in Vim. It is still not identical, though, since Vim only adds positions to the jump list when you use certain types of commands. For instance, if you use a "5k" (five lines upward) command in Vim, the new position replaces the previous one, while this plugin will keep the previous position as well (but not if you move just one line, as of this commit).